initialize bare objects passed as arguments
