+++ xnmt/attender.py
@@ -4,7 +4,8 @@ logger = logging.getLogger('xnmt')
 import math
 import dynet as dy
 
-from xnmt.serialize.serializable import Serializable, serializable_init, Ref, Path
+from xnmt.serialize.serializable import Serializable, Ref, Path
+from xnmt.serialize.serializer import serializable_init
 
 class Attender(object):
   '''
