+++ xnmt/batcher.py
@@ -3,7 +3,8 @@ import random
 import numpy as np
 import dynet as dy
 from xnmt.vocab import Vocab
-from xnmt.serialize.serializable import Serializable, serializable_init
+from xnmt.serialize.serializable import Serializable
+from xnmt.serialize.serializer import serializable_init
 
 class Batch(list):
   """
