+++ xnmt/bridge.py
@@ -1,7 +1,8 @@
 import dynet as dy
 
 import xnmt.linear
-from xnmt.serialize.serializable import Serializable, serializable_init, Ref, Path
+from xnmt.serialize.serializable import Serializable, Ref, Path
+from xnmt.serialize.serializer import serializable_init
 
 class Bridge(object):
   """
