+++ xnmt/lstm.py
@@ -6,7 +6,8 @@ import dynet as dy
 from xnmt.expression_sequence import ExpressionSequence, ReversedExpressionSequence
 from xnmt.events import register_xnmt_handler, handle_xnmt_event, register_xnmt_handler
 from xnmt.transducer import SeqTransducer, FinalTransducerState
-from xnmt.serialize.serializable import Serializable, serializable_init, Ref, Path
+from xnmt.serialize.serializable import Serializable, Ref, Path
+from xnmt.serialize.serializer import serializable_init
 
 class UniLSTMSeqTransducer(SeqTransducer, Serializable):
   """
