+++ xnmt/param_init.py
@@ -3,7 +3,8 @@ import math
 import numpy as np
 import dynet as dy
 
-from xnmt.serialize.serializable import Serializable, serializable_init
+from xnmt.serialize.serializable import Serializable
+from xnmt.serialize.serializer import serializable_init
 
 class ParamInitializer(object):
   """
