+++ xnmt/retriever.py
@@ -9,7 +9,8 @@ from simple_settings import settings
 import xnmt.batcher
 from xnmt.events import handle_xnmt_event
 from xnmt.generator import GeneratorModel
-from xnmt.serialize.serializable import Serializable, serializable_init
+from xnmt.serialize.serializable import Serializable
+from xnmt.serialize.serializer import serializable_init
 from xnmt.reports import Reportable
 from xnmt.expression_sequence import ExpressionSequence
