+++ xnmt/segmenting_composer.py
@@ -3,7 +3,8 @@ import dynet as dy
 import xnmt.linear
 import xnmt.embedder
 
-from xnmt.serialize.serializable import Serializable, Ref, Path, serializable_init
+from xnmt.serialize.serializable import Serializable, Ref, Path
+from xnmt.serialize.serializer import serializable_init
 from xnmt.events import register_xnmt_handler, handle_xnmt_event, register_xnmt_event
 from xnmt.reports import Reportable
 from xnmt.vocab import Vocab
